Renames Actions\Indexation namespace to Actions\Indexing.
