+++ src/actions/indexing/abstract-link-indexing-action.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yoast\WP\SEO\Actions\Indexation;
+namespace Yoast\WP\SEO\Actions\Indexing;
 
 use wpdb;
 use Yoast\WP\SEO\Builders\Indexable_Link_Builder;
