+++ src/actions/indexing/post-link-indexing-action.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yoast\WP\SEO\Actions\Indexation;
+namespace Yoast\WP\SEO\Actions\Indexing;
 
 use Yoast\WP\Lib\Model;
 use Yoast\WP\SEO\Helpers\Post_Type_Helper;
