+++ tests/unit/actions/indexation/indexable-post-type-archive-indexation-action-test.php
@@ -6,7 +6,7 @@ use Brain\Monkey;
 use Mockery;
 use Mockery\MockInterface;
 use Yoast\WP\Lib\ORM;
-use Yoast\WP\SEO\Actions\Indexation\Indexable_Post_Type_Archive_Indexation_Action;
+use Yoast\WP\SEO\Actions\Indexing\Indexable_Post_Type_Archive_Indexation_Action;
 use Yoast\WP\SEO\Builders\Indexable_Builder;
 use Yoast\WP\SEO\Helpers\Post_Type_Helper;
 use Yoast\WP\SEO\Repositories\Indexable_Repository;
@@ -19,7 +19,7 @@ use Yoast\WP\SEO\Tests\Unit\TestCase;
  * @group actions
  * @group indexing
  *
- * @coversDefaultClass \Yoast\WP\SEO\Actions\Indexation\Indexable_Post_Type_Archive_Indexation_Action
+ * @coversDefaultClass \Yoast\WP\SEO\Actions\Indexing\Indexable_Post_Type_Archive_Indexation_Action
  */
 class Indexable_Post_Type_Archive_Indexation_Action_Test extends TestCase {
