+++ tests/unit/actions/indexation/indexable-prepare-indexation-action-test.php
@@ -3,7 +3,7 @@
 namespace Yoast\WP\SEO\Tests\Unit\Actions\Indexation;
 
 use Mockery;
-use Yoast\WP\SEO\Actions\Indexation\Indexable_Prepare_Indexation_Action;
+use Yoast\WP\SEO\Actions\Indexing\Indexable_Prepare_Indexation_Action;
 use Yoast\WP\SEO\Helpers\Date_Helper;
 use Yoast\WP\SEO\Helpers\Options_Helper;
 use Yoast\WP\SEO\Tests\Unit\TestCase;
@@ -14,7 +14,7 @@ use Yoast\WP\SEO\Tests\Unit\TestCase;
  * @group actions
  * @group indexing
  *
- * @coversDefaultClass \Yoast\WP\SEO\Actions\Indexation\Indexable_Prepare_Indexation_Action
+ * @coversDefaultClass \Yoast\WP\SEO\Actions\Indexing\Indexable_Prepare_Indexation_Action
  */
 class Indexable_Prepare_Indexation_Action_Test extends TestCase {
